+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/schedule/ScheduleDataSourceImpl.kt
@@ -1,16 +1,16 @@
 package com.daedan.festabook.data.datasource.remote.schedule
 
-import com.daedan.festabook.data.api.ScheduleApi
-import com.daedan.festabook.data.datasource.remote.adapter.ApiResult
-import com.daedan.festabook.data.model.ScheduleDateResponse
-import com.daedan.festabook.data.model.ScheduleEventResponse
+import com.daedan.festabook.data.datasource.remote.ApiResult
+import com.daedan.festabook.data.model.response.ScheduleDateResponse
+import com.daedan.festabook.data.model.response.ScheduleEventResponse
+import com.daedan.festabook.data.service.ScheduleService
 
 class ScheduleDataSourceImpl(
-    private val scheduleApi: ScheduleApi,
+    private val scheduleService: ScheduleService,
 ) : ScheduleDataSource {
     override suspend fun fetchScheduleEventsById(eventDateId: Long): ApiResult<List<ScheduleEventResponse>> =
-        ApiResult.toApiResult { scheduleApi.fetchScheduleEventsById(eventDateId) }
+        ApiResult.toApiResult { scheduleService.fetchScheduleEventsById(eventDateId) }
 
     override suspend fun fetchScheduleDates(): ApiResult<List<ScheduleDateResponse>> =
-        ApiResult.toApiResult { scheduleApi.fetchScheduleDates() }
+        ApiResult.toApiResult { scheduleService.fetchScheduleDates() }
 }
