+++ android/app/src/main/java/com/daedan/festabook/data/repository/ScheduleRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.daedan.festabook.data.repository
 
-import com.daedan.festabook.data.datasource.remote.adapter.ApiResult
 import com.daedan.festabook.data.datasource.remote.schedule.ScheduleDataSource
-import com.daedan.festabook.data.model.toDomain
+import com.daedan.festabook.data.model.response.toDomain
+import com.daedan.festabook.data.util.toResult
 import com.daedan.festabook.domain.model.ScheduleDate
 import com.daedan.festabook.domain.model.ScheduleEvent
 import com.daedan.festabook.domain.repository.ScheduleRepository
@@ -20,12 +20,3 @@ class ScheduleRepositoryImpl(
         return response.map { scheduleEventResponses -> scheduleEventResponses.map { it.toDomain() } }
     }
 }
-
-fun <T> ApiResult<T>.toResult(): Result<T> =
-    when (this) {
-        is ApiResult.Success -> Result.success(data)
-        is ApiResult.ClientError -> Result.failure(Exception("Client error: $code $message"))
-        is ApiResult.ServerError -> Result.failure(Exception("Server error: $code $message"))
-        is ApiResult.NetworkError -> Result.failure(throwable)
-        is ApiResult.UnknownError -> Result.failure(Exception("Unknown error"))
-    }
