+++ android/app/src/main/java/com/daedan/festabook/data/service/api/ApiClient.kt
@@ -1,16 +1,14 @@
-package com.daedan.festabook.data.datasource.remote
+package com.daedan.festabook.data.service.api
 
 import com.daedan.festabook.BuildConfig
-import com.daedan.festabook.data.api.ScheduleApi
-import com.daedan.festabook.data.datasource.remote.adapter.ApiResultCallAdapterFactory
-import com.daedan.festabook.data.interceptor.FestaBookAuthInterceptor
+import com.daedan.festabook.data.service.ScheduleService
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 
-object NetworkModule {
+object ApiClient {
     private val json = Json { ignoreUnknownKeys = true }
 
     private val okHttpClient: OkHttpClient by lazy {
@@ -26,11 +24,10 @@ object NetworkModule {
             .baseUrl(BuildConfig.FESTABOOK_URL)
             .client(okHttpClient)
             .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
-            .addCallAdapterFactory(ApiResultCallAdapterFactory())
             .build()
     }
 
-    val scheduleApi: ScheduleApi by lazy {
-        retrofit.create(ScheduleApi::class.java)
+    val scheduleService: ScheduleService by lazy {
+        retrofit.create(ScheduleService::class.java)
     }
 }
